Replace tools.go w/ explicit versions in Makefile (#8)
Replace tools.go with explicit versioning of build-time
tooling in the Makefile (e.g. go run mytool@v1.2.3). Because this
is a library imported by other projects, my preference is to
keep go.mod narrowly focused on the dependencies of the code itself
and not the ancillary tools used by Elastic to maintain its NOTICE.txt
and license headers.

Co-authored-by: Panos Koutsovasilis <panos.koutsovasilis@elastic.co>